+++ src/Websites/Websites/Utilities/WebsitesClient.cs
@@ -53,7 +53,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             private set;
         }
 
-        public Site CreateWebApp(string resourceGroupName, string webAppName, string slotName, string location, string serverFarmId, CloningInfo cloningInfo, string aseName, string aseResourceGroupName, IDictionary<string, string> tags = null, ManagedServiceIdentity sourceIdentity=null)
+        public Site CreateWebApp(string resourceGroupName, string webAppName, string slotName, string location, string serverFarmId, CloningInfo cloningInfo, string aseName, string aseResourceGroupName, IDictionary<string, string> tags = null, ManagedServiceIdentity sourceIdentity = null)
         {
             Site createdWebSite = null;
             string qualifiedSiteName;
@@ -128,7 +128,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             string qualifiedSiteName;
             if (CmdletHelpers.ShouldUseDeploymentSlot(webAppName, slotName, out qualifiedSiteName))
             {
-                WrappedWebsitesClient.WebApps().CreateOrUpdateSlot(resourceGroupName, webAppName, webSiteToUpdate, slotName);
+                WrappedWebsitesClient.WebApps().CreateOrUpdateSlot(resourceGroupName, webAppName, slotName, webSiteToUpdate);
             }
             else
             {
@@ -161,10 +161,10 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                         else
                         {
                             WrappedWebsitesClient.WebApps().CreateOrUpdateHostNameBindingSlot(resourceGroupName, webAppName,
-                                hostName, new HostNameBinding
+                                hostName, slotName, new HostNameBinding
                                 {
                                     SiteName = webAppName,
-                                }, slotName);
+                                });
                         }
                     }
                 }
@@ -308,7 +308,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             };
 
             var publishingXml = (CmdletHelpers.ShouldUseDeploymentSlot(webSiteName, slotName, out qualifiedSiteName) ?
-                WrappedWebsitesClient.WebApps().ListPublishingProfileXmlWithSecretsSlot(resourceGroupName, webSiteName, options, slotName) :
+                WrappedWebsitesClient.WebApps().ListPublishingProfileXmlWithSecretsSlot(resourceGroupName, webSiteName, slotName, options) :
                 WrappedWebsitesClient.WebApps().ListPublishingProfileXmlWithSecrets(resourceGroupName, webSiteName, options));
             var doc = XDocument.Load(publishingXml, LoadOptions.None);
             if (outputFile != null)
@@ -345,7 +345,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             };
 
             var publishingXml = (CmdletHelpers.ShouldUseDeploymentSlot(webSiteName, slotName, out qualifiedSiteName) ?
-                WrappedWebsitesClient.WebApps().ListPublishingProfileXmlWithSecretsSlot(resourceGroupName, webSiteName, options, slotName) :
+                WrappedWebsitesClient.WebApps().ListPublishingProfileXmlWithSecretsSlot(resourceGroupName, webSiteName, slotName, options) :
                 WrappedWebsitesClient.WebApps().ListPublishingProfileXmlWithSecrets(resourceGroupName, webSiteName, options));
             var doc = XDocument.Load(publishingXml, LoadOptions.None);
             var profile = doc.Root == null ? null : doc.Root.Element("publishData") == null ? null : doc.Root.Element("publishData").Elements("publishProfile")
@@ -544,8 +544,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                     WrappedWebsitesClient.WebApps().UpdateConfigurationSlot(
                         resourceGroupName,
                         webSiteName,
-                        siteConfig.ConvertToSiteConfigResource(),
-                        slotName);
+                        slotName,
+                        siteConfig.ConvertToSiteConfigResource());
                 }
 
                 if (appSettings != null)
@@ -553,8 +553,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                     WrappedWebsitesClient.WebApps().UpdateApplicationSettingsSlot(
                         resourceGroupName,
                         webSiteName,
-                        new StringDictionary { Properties = appSettings },
-                        slotName);
+                        slotName,
+                        new StringDictionary { Properties = appSettings });
                 }
 
                 if (connectionStrings != null)
@@ -562,8 +562,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                     WrappedWebsitesClient.WebApps().UpdateConnectionStringsSlot(
                         resourceGroupName,
                         webSiteName,
-                        new ConnectionStringDictionary { Properties = connectionStrings },
-                        slotName);
+                        slotName,
+                        new ConnectionStringDictionary { Properties = connectionStrings });
                 }
 
                 if (azureStorageSettings != null)
@@ -571,8 +571,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                     WrappedWebsitesClient.WebApps().UpdateAzureStorageAccountsSlot(
                         resourceGroupName,
                         webSiteName,
-                        azureStorageSettings,
-                        slotName);
+                        slotName,
+                        azureStorageSettings);
                 }
             }
             else
@@ -704,8 +704,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                 return WrappedWebsitesClient.WebApps().UpdateBackupConfigurationSlot(
                     resourceGroupName,
                     webSiteName,
-                    newSchedule,
-                    slotName);
+                    slotName,
+                    newSchedule);
             }
             else
             {
@@ -723,7 +723,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             var useSlot = CmdletHelpers.ShouldUseDeploymentSlot(webSiteName, slotName, out qualifiedSiteName);
             if (useSlot)
             {
-                var backup = WrappedWebsitesClient.WebApps().BackupSlot(resourceGroupName, webSiteName, request, slotName);
+                var backup = WrappedWebsitesClient.WebApps().BackupSlot(resourceGroupName, webSiteName, slotName, request);
                 return backup;
             }
             else
@@ -821,8 +821,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                     resourceGroupName,
                     webSiteName,
                     backupId,
-                    request,
-                    slotName);
+                    slotName,
+                    request);
             }
             else
             {
@@ -867,7 +867,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             bool useSlot = CmdletHelpers.ShouldUseDeploymentSlot(webSiteName, slotName, out qualifiedSiteName);
             if (useSlot)
             {
-                WrappedWebsitesClient.WebApps().RestoreSnapshotSlot(resourceGroupName, webSiteName, restoreReq, slotName);
+                WrappedWebsitesClient.WebApps().RestoreSnapshotSlot(resourceGroupName, webSiteName, slotName, restoreReq);
             }
             else
             {
@@ -899,7 +899,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             bool useSlot = CmdletHelpers.ShouldUseDeploymentSlot(webSiteName, slotName, out qualifiedSiteName);
             if (useSlot)
             {
-                WrappedWebsitesClient.WebApps().BeginRestoreFromDeletedAppSlot(resourceGroupName, webSiteName, restoreReq, slotName);
+                WrappedWebsitesClient.WebApps().BeginRestoreFromDeletedAppSlot(resourceGroupName, webSiteName, slotName, restoreReq);
             }
             else
             {
@@ -1003,7 +1003,7 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
                 csmSlotEntity.PreserveVnet = preserveVnet.Value;
             }
 
-            WrappedWebsitesClient.WebApps().SwapSlotSlot(
+            WrappedWebsitesClient.WebApps().SwapSlotsSlot(
                 resourceGroupName,
                 webSiteName,
                 csmSlotEntity,
@@ -1022,8 +1022,8 @@ namespace Microsoft.Azure.Commands.WebApps.Utilities
             WrappedWebsitesClient.WebApps().ApplySlotConfigurationSlot(
                 resourceGroupName,
                 webSiteName,
-                csmSlotEntity,
-                sourceSlotName);
+                sourceSlotName,
+                csmSlotEntity);
         }
 
         public void SwapSlotWithPreviewResetSlotSwap(string resourceGroupName, string webSiteName, string sourceSlotName)